remove diff to master; undo renaming
