+++ digicampipe/calib/camera/random_triggers.py
